Moved `LocationDTO` into `Events/` folder
