+++ src/main/java/com/danielagapov/spawn/DTOs/Event/EventCreationDTO.java
@@ -1,6 +1,5 @@
 package com.danielagapov.spawn.DTOs.Event;
 
-import com.danielagapov.spawn.DTOs.LocationDTO;
 import lombok.Getter;
 import lombok.Setter;
