+++ src/main/java/com/danielagapov/spawn/DTOs/Event/LocationDTO.java
@@ -1,4 +1,4 @@
-package com.danielagapov.spawn.DTOs;
+package com.danielagapov.spawn.DTOs.Event;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
